PUBDEV-3200 - fixed String/Enum parsing issue + minor cleanup.
There was an incorrect handling of repeated values.
The assumption was that the values are repeated if the length for the field is marked as 0, which is not true.
Values are repeated if ColumnVector.isRepeating is set to true.
Most likely applies to all other columns which should get changed too.